# crates/tmf-apis/tmf654_prepay_balance/src/models.rs
//! TMF654 Prepay Balance models

use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use tmf_apis_core::BaseEntity;
use utoipa::ToSchema;
use uuid::Uuid;

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
pub enum BalanceType {
    Monetary,
    Data,
    Voice,
    Sms,
    Other,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
pub struct Money {
    pub value: f64,
    pub unit: String,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
pub struct PrepayBalance {
    #[serde(flatten)]
    pub base: BaseEntity,
    pub balance_type: BalanceType,
    pub remaining_value: Money,
    #[serde(skip_serializing_if = "Option::is_none")]
    #[schema(value_type = String, format = "uuid")]
    pub party_id: Option<Uuid>,
    #[serde(skip_serializing_if = "Option::is_none")]
    #[schema(value_type = String, format = "date-time")]
    pub valid_for_end: Option<DateTime<Utc>>,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
pub struct CreatePrepayBalanceRequest {
    pub name: String,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub description: Option<String>,
    pub balance_type: BalanceType,
    pub remaining_value: Money,
    #[serde(skip_serializing_if = "Option::is_none")]
    #[schema(value_type = String, format = "uuid")]
    pub party_id: Option<Uuid>,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
pub struct AdjustBalanceRequest {
    /// Positive credits, negative debits (same unit as balance).
    pub delta: Money,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub reason: Option<String>,
}

#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
pub struct UpdatePrepayBalanceRequest {
    #[serde(skip_serializing_if = "Option::is_none")]
    pub remaining_value: Option<Money>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub description: Option<String>,
}