# openapi/tmf-apis-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BSS/OSS Rust - TM Forum Open APIs",
    "description": "TM Forum Open API implementation for BSS/OSS ecosystem (TMF620, TMF622, TMF637, TMF629, TMF678, TMF677, TMF646, TMF641, TMF638, TMF640, TMF702, TMF639, TMF645, TMF635, TMF668, TMF632, TMF669, TMF642, TMF656, TMF621, TMF648, TMF633, TMF634, TMF679, TMF666, TMF676, TMF651, TMF654)",
    "contact": {
      "name": "Roberto de Souza",
      "email": "rabbittrix@hotmail.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.3.2"
  },
  "paths": {
...
    "/tmf-api/prepayBalanceManagement/v4/prepayBalance": {
      "get": {
        "tags": [
          "TMF654"
        ],
        "operationId": "get_balances",
        "responses": {
          "200": {
            "description": "List of prepay balances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrepayBalance"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TMF654"
        ],
        "operationId": "create_balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrepayBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Balance created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepayBalance"
                }
              }
            }
          }
        }
      }
    },
    "/tmf-api/prepayBalanceManagement/v4/prepayBalance/{id}": {
      "get": {
        "tags": [
          "TMF654"
        ],
        "operationId": "get_balance_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Balance ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepayBalance"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TMF654"
        ],
        "operationId": "delete_balance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Balance ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      },
      "patch": {
        "tags": [
          "TMF654"
        ],
        "operationId": "update_balance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Balance ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrepayBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Balance updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepayBalance"
                }
              }
            }
          }
        }
      }
    },
    "/tmf-api/prepayBalanceManagement/v4/prepayBalance/{id}/adjust": {
      "post": {
        "tags": [
          "TMF654"
        ],
        "operationId": "adjust_balance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Balance ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Balance adjusted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepayBalance"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdjustBalanceRequest": {
        "type": "object",
        "required": [
          "delta"
        ],
        "properties": {
          "delta": {
            "$ref": "#/components/schemas/Money"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "BalanceType": {
        "type": "string",
        "enum": [
          "MONETARY",
          "DATA",
          "VOICE",
          "SMS",
          "OTHER"
        ]
      },
      "BaseEntity": {
        "type": "object",
        "description": "Base entity with common fields",
        "required": [
          "id",
          "name",
          "lifecycle_status"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Description",
            "nullable": true
          },
          "href": {
            "type": "string",
            "description": "Resource URI",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "last_update": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date"
          },
          "lifecycle_status": {
            "$ref": "#/components/schemas/LifecycleStatus"
          },
          "name": {
            "type": "string",
            "description": "Name"
          },
          "valid_for": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimePeriod"
              }
            ],
            "nullable": true
          },
          "version": {
            "type": "string",
            "description": "Version",
            "nullable": true
          }
        }
      },
...
      "CreatePrepayBalanceRequest": {
        "type": "object",
        "required": [
          "name",
          "balance_type",
          "remaining_value"
        ],
        "properties": {
          "balance_type": {
            "$ref": "#/components/schemas/BalanceType"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "party_id": {
            "type": "string",
            "format": "uuid"
          },
          "remaining_value": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
...
      "LifecycleStatus": {
        "type": "string",
        "description": "Lifecycle status of a resource",
        "enum": [
          "IN_STUDY",
          "IN_DESIGN",
          "IN_TEST",
          "ACTIVE",
          "LAUNCHED",
          "RETIRED",
          "OBSOLETE",
          "REJECTED"
        ]
      },
      "Money": {
        "type": "object",
        "required": [
          "value",
          "unit"
        ],
        "properties": {
          "unit": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "PrepayBalance": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "required": [
              "balance_type",
              "remaining_value"
            ],
            "properties": {
              "balance_type": {
                "$ref": "#/components/schemas/BalanceType"
              },
              "party_id": {
                "type": "string",
                "format": "uuid"
              },
              "remaining_value": {
                "$ref": "#/components/schemas/Money"
              },
              "valid_for_end": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
      "TimePeriod": {
        "type": "object",
        "description": "Time period for validity",
        "required": [
          "start_date_time"
        ],
        "properties": {
          "end_date_time": {
            "type": "string",
            "format": "date-time"
          },
          "start_date_time": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdatePrepayBalanceRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "remaining_value": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "nullable": true
          }
        }
      },
...
    }
  },
...
}